feat: tools refactor
