+++ src/tools/dashboards/get.ts
@@ -1,0 +1,33 @@
+import { z } from "zod";
+import type { Context, Tool } from "../types";
+
+const schema = z.object({
+    dashboardId: z.number(),
+});
+
+type Params = z.infer<typeof schema>;
+
+export const getHandler = async (context: Context, params: Params) => {
+    const { dashboardId } = params;
+    const projectId = await context.getProjectId();
+    const dashboardResult = await context.api
+        .dashboards({ projectId })
+        .get({ dashboardId });
+
+    if (!dashboardResult.success) {
+        throw new Error(`Failed to get dashboard: ${dashboardResult.error.message}`);
+    }
+
+    return { content: [{ type: "text", text: JSON.stringify(dashboardResult.data) }] };
+};
+
+const tool = (): Tool<typeof schema> => ({
+    name: "dashboard-get",
+    description: `
+        - Get a specific dashboard by ID.
+    `,
+    schema,
+    handler: getHandler,
+});
+
+export default tool;
