+++ src/tools/errorTracking/errorDetails.ts
@@ -1,0 +1,48 @@
+import { z } from "zod";
+import type { Context, Tool } from "../types";
+import { ErrorDetailsSchema } from "../../schema/errors";
+
+const schema = z.object({
+    data: ErrorDetailsSchema,
+});
+
+type Params = z.infer<typeof schema>;
+
+export const errorDetailsHandler = async (context: Context, params: Params) => {
+    const { data } = params;
+    const projectId = await context.getProjectId();
+
+    const errorQuery = {
+        kind: "ErrorTrackingQuery",
+        dateRange: {
+            date_from:
+                data.dateFrom?.toISOString() ||
+                new Date(Date.now() - 7 * 24 * 60 * 60 * 1000).toISOString(),
+            date_to: data.dateTo?.toISOString() || new Date().toISOString(),
+        },
+        volumeResolution: 0,
+        issueId: data.issueId,
+    };
+
+    const errorsResult = await context.api
+        .query({ projectId })
+        .execute({ queryBody: errorQuery });
+    if (!errorsResult.success) {
+        throw new Error(`Failed to get error details: ${errorsResult.error.message}`);
+    }
+
+    return {
+        content: [{ type: "text", text: JSON.stringify(errorsResult.data.results) }],
+    };
+};
+
+const tool = (): Tool<typeof schema> => ({
+    name: "error-details",
+    description: `
+        - Use this tool to get the details of an error in the project.
+    `,
+    schema,
+    handler: errorDetailsHandler,
+});
+
+export default tool;
