+++ src/tools/types.ts
@@ -1,0 +1,27 @@
+import { z } from "zod";
+import { ApiClient } from "../api/client";
+import { DurableObjectCache } from "../lib/utils/cache/DurableObjectCache";
+
+export type State = {
+    projectId: string | undefined;
+    orgId: string | undefined;
+    distinctId: string | undefined;
+};
+
+export type Context = {
+    api: ApiClient;
+    cache: DurableObjectCache<State>;
+    env: Env;
+    getProjectId: () => Promise<string>;
+    getOrgID: () => Promise<string>;
+    getDistinctId: () => Promise<string>;
+};
+
+export type Tool<TSchema extends z.ZodTypeAny = z.ZodTypeAny> = {
+    name: string;
+    description: string;
+    schema: TSchema;
+    handler: (context: Context, params: z.infer<TSchema>) => Promise<any>;
+};
+
+export type ZodObjectAny = z.ZodObject<any, any, any, any, any>;
